+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Play/Pause/Reset controls
 - Canvas screenshot
 - Resolution and aspect ratio controls
+- Display fragment shader with line numbers on error
 
 ### Changed
 - `shader-workshop` program has been renamed to `sw-server`